Finally fixed Josep's mess

+++ submesh.h
@@ -36,7 +36,7 @@ class SubMesh :
 public:
     SubMesh();
     SubMesh(VertexFormat vertexFormat, void *data, int size);
-    SubMesh(VertexFormat vertexFormat, unsigned int *data, int size, unsigned int *indices, int indices_count);
+    SubMesh(VertexFormat vertexFormat, float *data, int size, unsigned int *indices, int indices_count);
     ~SubMesh();
 
     void update();
@@ -45,7 +45,7 @@ public:
 
 private:
     VertexFormat vertexFormat;
-    unsigned int *data = nullptr;
+    float *data = nullptr;
     size_t data_size = 0;
     unsigned int *indices = nullptr;
     size_t indices_count = 0;
